src/test/java/io/github/vlad324/n430/SolutionTest.java
package io.github.vlad324.n430;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

class SolutionTest {
    private final Solution solution = new Solution();

    private static Stream<Arguments> testDataProvider() {
        return Stream.of(
            arguments()
        );
    }

    @ParameterizedTest
    @MethodSource("testDataProvider")
    void should_produce_expected_result(Solution.Node head, Solution.Node expected) {
        // when
        final var actual = solution.flatten(head);

        // then
        assertThat(actual).isEqualTo(expected);
    }
}